refactor(gateway): give talk.client.create its own module
talk-client.ts sat just under the 700-line cap with a ~360-line create
handler inline. Move it to talk-client-create.ts, matching the existing
sibling handler modules (talk-client-run-ownership, talk-session-mark), so
the handlers object reads as a routing table again.

Extraction exposed a redundant type assertion; removing it shrinks the
assertion-safety baseline by one entry.
